+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/agent/Agent.java
@@ -1,0 +1,344 @@
+package datadog.trace.agent;
+
+import java.lang.instrument.Instrumentation;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.net.URL;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Agent start up logic.
+ *
+ * <p>This class is loaded and called by {@code datadog.trace.agent.AgentBootstrap}
+ *
+ * <p>The intention is for this class to be loaded by bootstrap classloader to make sure we have
+ * unimpeded access to the rest of Datadog's agent parts.
+ */
+public class Agent {
+
+  private static final String SIMPLE_LOGGER_SHOW_DATE_TIME_PROPERTY =
+      "datadog.slf4j.simpleLogger.showDateTime";
+  private static final String SIMPLE_LOGGER_DATE_TIME_FORMAT_PROPERTY =
+      "datadog.slf4j.simpleLogger.dateTimeFormat";
+  private static final String SIMPLE_LOGGER_DATE_TIME_FORMAT_DEFAULT =
+      "'[dd.trace 'yyyy-MM-dd HH:mm:ss:SSS Z']'";
+  private static final String SIMPLE_LOGGER_DEFAULT_LOG_LEVEL_PROPERTY =
+      "datadog.slf4j.simpleLogger.defaultLogLevel";
+
+  // We cannot use lombok here because we need to configure logger first
+  private static final Logger LOGGER;
+
+  static {
+    // We can configure logger here because datadog.trace.agent.AgentBootstrap doesn't touch it.
+    configureLogger();
+    LOGGER = LoggerFactory.getLogger(Agent.class);
+  }
+
+  // fields must be managed under class lock
+  private static ClassLoader AGENT_CLASSLOADER = null;
+  private static ClassLoader JMXFETCH_CLASSLOADER = null;
+
+  public static void start(final Instrumentation inst, final URL bootstrapURL) throws Exception {
+    startDatadogAgent(inst, bootstrapURL);
+
+    final boolean appUsingCustomLogManager = isAppUsingCustomLogManager();
+
+    /*
+     * java.util.logging.LogManager maintains a final static LogManager, which is created during class initialization.
+     *
+     * JMXFetch uses jre bootstrap classes which touch this class. This means applications which require a custom log
+     * manager may not have a chance to set the global log manager if jmxfetch runs first. JMXFetch will incorrectly
+     * set the global log manager in cases where the app sets the log manager system property or when the log manager
+     * class is not on the system classpath.
+     *
+     * Our solution is to delay the initialization of jmxfetch when we detect a custom log manager being used.
+     *
+     * Once we see the LogManager class loading, it's safe to start jmxfetch because the application is already setting
+     * the global log manager and jmxfetch won't be able to touch it due to classloader locking.
+     */
+    if (appUsingCustomLogManager) {
+      LOGGER.debug("Custom logger detected. Delaying JMXFetch initialization.");
+      registerLogManagerCallback(new StartJmxFetchCallback(inst, bootstrapURL));
+    } else {
+      startJmxFetch(inst, bootstrapURL);
+    }
+
+    /*
+     * Similar thing happens with DatadogTracer on (at least) zulu-8 because it uses OkHttp which indirectly loads JFR
+     * events which in turn loads LogManager. This is not a problem on newer JDKs because there JFR uses different
+     * logging facility.
+     */
+    if (isJavaBefore9() && appUsingCustomLogManager) {
+      LOGGER.debug("Custom logger detected. Delaying Datadog Tracer initialization.");
+      registerLogManagerCallback(new InstallDatadogTracerCallback(inst, bootstrapURL));
+    } else {
+      installDatadogTracer(inst, bootstrapURL);
+    }
+  }
+
+  private static void registerLogManagerCallback(final Runnable callback) throws Exception {
+    final Class<?> agentInstallerClass =
+        AGENT_CLASSLOADER.loadClass("datadog.trace.agent.tooling.AgentInstaller");
+    final Method registerCallbackMethod =
+        agentInstallerClass.getMethod("registerClassLoadCallback", String.class, Runnable.class);
+    registerCallbackMethod.invoke(null, "java.util.logging.LogManager", callback);
+  }
+
+  protected abstract static class ClassLoadCallBack implements Runnable {
+
+    final Instrumentation inst;
+    final URL bootstrapURL;
+
+    ClassLoadCallBack(final Instrumentation inst, final URL bootstrapURL) {
+      this.inst = inst;
+      this.bootstrapURL = bootstrapURL;
+    }
+
+    @Override
+    public void run() {
+      /*
+       * This callback is called from within bytecode transformer. This can be a problem if callback tries
+       * to load classes being transformed. To avoid this we start a thread here that calls the callback.
+       * This seems to resolve this problem.
+       */
+      final Thread thread =
+          new Thread(
+              new Runnable() {
+                @Override
+                public void run() {
+                  try {
+                    execute();
+                  } catch (final Exception e) {
+                    LOGGER.error("Failed to run class loader callback {}", getName(), e);
+                  }
+                }
+              });
+      thread.setName("dd-agent-startup-" + getName());
+      thread.setDaemon(true);
+      thread.start();
+    }
+
+    public abstract String getName();
+
+    public abstract void execute() throws Exception;
+  }
+
+  protected static class StartJmxFetchCallback extends ClassLoadCallBack {
+    StartJmxFetchCallback(final Instrumentation inst, final URL bootstrapURL) {
+      super(inst, bootstrapURL);
+    }
+
+    @Override
+    public String getName() {
+      return "jmxfetch";
+    }
+
+    @Override
+    public void execute() throws Exception {
+      startJmxFetch(inst, bootstrapURL);
+    }
+  }
+
+  protected static class InstallDatadogTracerCallback extends ClassLoadCallBack {
+    InstallDatadogTracerCallback(final Instrumentation inst, final URL bootstrapURL) {
+      super(inst, bootstrapURL);
+    }
+
+    @Override
+    public String getName() {
+      return "datadog-tracer";
+    }
+
+    @Override
+    public void execute() throws Exception {
+      installDatadogTracer(inst, bootstrapURL);
+    }
+  }
+
+  private static synchronized void startDatadogAgent(
+      final Instrumentation inst, final URL bootstrapURL) throws Exception {
+    if (AGENT_CLASSLOADER == null) {
+      final ClassLoader agentClassLoader =
+          createDatadogClassLoader("agent-tooling-and-instrumentation.isolated", bootstrapURL);
+      final ClassLoader contextLoader = Thread.currentThread().getContextClassLoader();
+      try {
+        Thread.currentThread().setContextClassLoader(agentClassLoader);
+        final Class<?> agentInstallerClass =
+            agentClassLoader.loadClass("datadog.trace.agent.tooling.AgentInstaller");
+        final Method agentInstallerMethod =
+            agentInstallerClass.getMethod("installBytebuddyAgent", Instrumentation.class);
+        agentInstallerMethod.invoke(null, inst);
+        AGENT_CLASSLOADER = agentClassLoader;
+      } finally {
+        Thread.currentThread().setContextClassLoader(contextLoader);
+      }
+    }
+  }
+
+  private static synchronized void installDatadogTracer(
+      final Instrumentation inst, final URL bootstrapURL) throws Exception {
+    if (AGENT_CLASSLOADER == null) {
+      throw new IllegalStateException("Datadog agent should have been started already");
+    }
+    final ClassLoader contextLoader = Thread.currentThread().getContextClassLoader();
+    // TracerInstaller.installGlobalTracer can be called multiple times without any problem
+    // so there is no need to have a 'datadogTracerInstalled' flag here.
+    try {
+      Thread.currentThread().setContextClassLoader(AGENT_CLASSLOADER);
+      // install global tracer
+      final Class<?> tracerInstallerClass =
+          AGENT_CLASSLOADER.loadClass("datadog.trace.agent.tooling.TracerInstaller");
+      final Method tracerInstallerMethod = tracerInstallerClass.getMethod("installGlobalTracer");
+      tracerInstallerMethod.invoke(null);
+      final Method logVersionInfoMethod = tracerInstallerClass.getMethod("logVersionInfo");
+      logVersionInfoMethod.invoke(null);
+    } finally {
+      Thread.currentThread().setContextClassLoader(contextLoader);
+    }
+  }
+
+  private static synchronized void startJmxFetch(final Instrumentation inst, final URL bootstrapURL)
+      throws Exception {
+    if (JMXFETCH_CLASSLOADER == null) {
+      final ClassLoader jmxFetchClassLoader =
+          createDatadogClassLoader("agent-jmxfetch.isolated", bootstrapURL);
+      final ClassLoader contextLoader = Thread.currentThread().getContextClassLoader();
+      try {
+        Thread.currentThread().setContextClassLoader(jmxFetchClassLoader);
+        final Class<?> jmxFetchAgentClass =
+            jmxFetchClassLoader.loadClass("datadog.trace.agent.jmxfetch.JMXFetch");
+        final Method jmxFetchInstallerMethod = jmxFetchAgentClass.getMethod("run");
+        jmxFetchInstallerMethod.invoke(null);
+        JMXFETCH_CLASSLOADER = jmxFetchClassLoader;
+      } finally {
+        Thread.currentThread().setContextClassLoader(contextLoader);
+      }
+    }
+  }
+
+  private static void configureLogger() {
+    setSystemPropertyDefault(SIMPLE_LOGGER_SHOW_DATE_TIME_PROPERTY, "true");
+    setSystemPropertyDefault(
+        SIMPLE_LOGGER_DATE_TIME_FORMAT_PROPERTY, SIMPLE_LOGGER_DATE_TIME_FORMAT_DEFAULT);
+
+    if (isDebugMode()) {
+      setSystemPropertyDefault(SIMPLE_LOGGER_DEFAULT_LOG_LEVEL_PROPERTY, "DEBUG");
+    }
+  }
+
+  private static void setSystemPropertyDefault(final String property, final String value) {
+    if (System.getProperty(property) == null) {
+      System.setProperty(property, value);
+    }
+  }
+
+  /**
+   * Create the datadog classloader. This must be called after the bootstrap jar has been appened to
+   * the bootstrap classpath.
+   *
+   * @param innerJarFilename Filename of internal jar to use for the classpath of the datadog
+   *     classloader
+   * @param bootstrapURL
+   * @return Datadog Classloader
+   */
+  private static ClassLoader createDatadogClassLoader(
+      final String innerJarFilename, final URL bootstrapURL) throws Exception {
+    final ClassLoader agentParent;
+    if (isJavaBefore9()) {
+      agentParent = null; // bootstrap
+    } else {
+      // platform classloader is parent of system in java 9+
+      agentParent = getPlatformClassLoader();
+    }
+
+    final Class<?> loaderClass =
+        ClassLoader.getSystemClassLoader().loadClass("datadog.trace.bootstrap.DatadogClassLoader");
+    final Constructor constructor =
+        loaderClass.getDeclaredConstructor(URL.class, String.class, ClassLoader.class);
+    return (ClassLoader) constructor.newInstance(bootstrapURL, innerJarFilename, agentParent);
+  }
+
+  private static ClassLoader getPlatformClassLoader()
+      throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
+    /*
+     Must invoke ClassLoader.getPlatformClassLoader by reflection to remain
+     compatible with java 7 + 8.
+    */
+    final Method method = ClassLoader.class.getDeclaredMethod("getPlatformClassLoader");
+    return (ClassLoader) method.invoke(null);
+  }
+
+  /**
+   * Determine if we should log in debug level according to dd.trace.debug
+   *
+   * @return true if we should
+   */
+  private static boolean isDebugMode() {
+    final String tracerDebugLevelSysprop = "dd.trace.debug";
+    final String tracerDebugLevelProp = System.getProperty(tracerDebugLevelSysprop);
+
+    if (tracerDebugLevelProp != null) {
+      return Boolean.parseBoolean(tracerDebugLevelProp);
+    }
+
+    final String tracerDebugLevelEnv =
+        System.getenv(tracerDebugLevelSysprop.replace('.', '_').toUpperCase());
+
+    if (tracerDebugLevelEnv != null) {
+      return Boolean.parseBoolean(tracerDebugLevelEnv);
+    }
+    return false;
+  }
+
+  /**
+   * Search for java or datadog-tracer sysprops which indicate that a custom log manager will be
+   * used. Also search for any app classes known to set a custom log manager.
+   *
+   * @return true if we detect a custom log manager being used.
+   */
+  private static boolean isAppUsingCustomLogManager() {
+    final String tracerCustomLogManSysprop = "dd.app.customlogmanager";
+    final String customLogManagerProp = System.getProperty(tracerCustomLogManSysprop);
+    final String customLogManagerEnv =
+        System.getenv(tracerCustomLogManSysprop.replace('.', '_').toUpperCase());
+
+    if (customLogManagerProp != null || customLogManagerEnv != null) {
+      LOGGER.debug("Prop - customlogmanager: " + customLogManagerProp);
+      LOGGER.debug("Env - customlogmanager: " + customLogManagerEnv);
+      // Allow setting to skip these automatic checks:
+      return Boolean.parseBoolean(customLogManagerProp)
+          || Boolean.parseBoolean(customLogManagerEnv);
+    }
+
+    final String jbossHome = System.getenv("JBOSS_HOME");
+    if (jbossHome != null) {
+      LOGGER.debug("Env - jboss: " + jbossHome);
+      // JBoss/Wildfly is known to set a custom log manager after startup.
+      // Originally we were checking for the presence of a jboss class,
+      // but it seems some non-jboss applications have jboss classes on the classpath.
+      // This would cause jmxfetch initialization to be delayed indefinitely.
+      // Checking for an environment variable required by jboss instead.
+      return true;
+    }
+
+    final String logManagerProp = System.getProperty("java.util.logging.manager");
+    if (logManagerProp != null) {
+      final boolean onSysClasspath =
+          ClassLoader.getSystemResource(logManagerProp.replaceAll("\\.", "/") + ".class") != null;
+      LOGGER.debug("Prop - logging.manager: " + logManagerProp);
+      LOGGER.debug("logging.manager on system classpath: " + onSysClasspath);
+      // Some applications set java.util.logging.manager but never actually initialize the logger.
+      // Check to see if the configured manager is on the system classpath.
+      // If so, it should be safe to initialize jmxfetch which will setup the log manager.
+      return !onSysClasspath;
+    }
+
+    return false;
+  }
+
+  private static boolean isJavaBefore9() {
+    return System.getProperty("java.version").startsWith("1.");
+  }
+}
