+++ dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/Constants.java
@@ -15,7 +15,7 @@ public final class Constants {
    */
   public static final String[] BOOTSTRAP_PACKAGE_PREFIXES = {
     "datadog.slf4j",
-    "datadog.trace.agent.TracingAgent",
+    "datadog.trace.agent",
     "datadog.trace.api",
     "datadog.trace.bootstrap",
     "datadog.trace.context",
