+++ dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/context/FieldBackedProvider.java
@@ -73,8 +73,8 @@ public class FieldBackedProvider implements InstrumentationContextProvider {
 
   /**
    * Note: the value here has to be inside on of the prefixes in
-   * datadog.trace.agent.tooling.Utils#BOOTSTRAP_PACKAGE_PREFIXES. This ensures that 'isolating' (or
-   * 'module') classloaders like jboss and osgi see injected classes. This works because we
+   * datadog.trace.agent.tooling.Constants#BOOTSTRAP_PACKAGE_PREFIXES. This ensures that 'isolating'
+   * (or 'module') classloaders like jboss and osgi see injected classes. This works because we
    * instrument those classloaders to load everything inside bootstrap packages.
    */
   private static final String DYNAMIC_CLASSES_PACKAGE =
