+++ dd-java-agent/src/test/java/jvmbootstraptest/AgentLoadedChecker.java
@@ -7,12 +7,11 @@ public class AgentLoadedChecker {
   public static void main(final String[] args) throws ClassNotFoundException {
     // Empty classloader that delegates to bootstrap
     final URLClassLoader emptyClassLoader = new URLClassLoader(new URL[] {}, null);
-    final Class agentClass = emptyClassLoader.loadClass("datadog.trace.agent.TracingAgent");
+    final Class agentClass = emptyClassLoader.loadClass("datadog.trace.agent.Agent");
 
     if (agentClass.getClassLoader() != null) {
       throw new RuntimeException(
-          "TracingAgent loaded into classloader other than bootstrap: "
-              + agentClass.getClassLoader());
+          "Agent loaded into classloader other than bootstrap: " + agentClass.getClassLoader());
     }
   }
 }
