+++ dd-java-agent/testing/src/main/groovy/datadog/trace/agent/test/SpockRunner.java
@@ -32,7 +32,7 @@ public class SpockRunner extends Sputnik {
    */
   public static final String[] BOOTSTRAP_PACKAGE_PREFIXES_COPY = {
     "datadog.slf4j",
-    "datadog.trace.agent.TracingAgent",
+    "datadog.trace.agent",
     "datadog.trace.api",
     "datadog.trace.bootstrap",
     "datadog.trace.context",
